+++ lib/rpush/client/mongoid/apns/feedback.rb
@@ -13,6 +13,9 @@ module Rpush
           validates :device_token, presence: true
           validates :failed_at, presence: true
 
+          index device_token: 1
+          index app_id: 1
+
           validates_with Rpush::Client::ActiveModel::Apns::DeviceTokenFormatValidator
         end
       end