+++ lib/rpush/client/mongoid/notification.rb
@@ -37,6 +37,11 @@ module Rpush
         increments :integer_id, model_name: name
         index integer_id: 1
 
+        index delivered: 1, failed: 1, deliver_after: 1, processing: 1
+        index delivered: 1, failed: 1
+        index device_token: 1
+        index app_id: 1
+
         belongs_to :app
       end
     end