+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * ActiveRecord logging is no longer redirected to rpush.log when embedded (#138).
   * Support for WNS (Windows RT) (#137).
   * Indexes added to some Mongoid fields (#151).
+  * Added support for Oracle.
 
   Bug fixes:
   * Fix for handling APNs error when using `rpush push` or `Rpush.push`.